[CoreBundle] fixes forms and data types
