+++ main/app/Resources/modules/data/ip/index.js
@@ -1,15 +1,21 @@
-import isEmpty from 'lodash/isEmpty'
-import {chain, array, ip, notBlank} from '#/main/core/validation'
+import {trans} from '#/main/app/intl/translation'
 
-import {IpGroup} from '#/main/core/layout/form/components/group/ip-group'
+import {ip} from '#/main/app/data/ip/validators'
+import {IpGroup} from '#/main/app/data/ip/components/group'
 
 // TODO : implement IP v6 input
 
 const dataType = {
   name: 'ip',
 
+  meta: {
+    creatable: false,
+    label: trans('ip', {}, 'data'),
+    description: trans('ip_desc', {}, 'data')
+  },
+
   /**
-   * Validates an IP string or IPs list.
+   * Validates an IP string.
    *   - it MUST contains 4 groups separated by ".".
    *   - each group MUST be a number between 0 and 255 or "*".
    *
@@ -18,28 +24,7 @@ const dataType = {
    *
    * @return {boolean}
    */
-  validate: (value, options = {}) => {
-    if (options.multiple) {
-      return chain(value, options, [array, (value) => {
-        if (value) {
-          const errors = {}
-
-          value.map((ipValue, index) => {
-            const error = chain(ipValue, {}, [notBlank, ip])
-            if (error) {
-              errors[index] = error
-            }
-          })
-
-          if (!isEmpty(errors)) {
-            return errors
-          }
-        }
-      }])
-    } else {
-      return ip(value)
-    }
-  },
+  validate: (value, options = {}) => ip(value, options),
 
   components: {
     form: IpGroup
