+++ main/core/Resources/less/layout/form/fields/index.less
@@ -24,22 +24,6 @@ input[type=checkbox] {
     cursor: pointer;
 }
 
-.checkbox-with-input,
-.radio-with-input {
-    line-height: 34px;
-    min-height: 40px;
-
-    .radio,
-    .checkbox {
-        display: inline-block;
-        margin: 0;
-    }
-
-    input[type=radio] {
-        margin-top: 10px;
-    }
-}
-
 // Fixes bootstrap that do not remove label font-weight like `.radio` and `.checkbox`
 // And so do not work as expected
 .radio-inline,
