--- main/core/Resources/less/libraries/index.less
@@ -1,6 +1,0 @@
-/*
- * Libraries barrel
- */
-
-@import "jquery-ui";
-@import "video-js";
