--- main/core/Resources/less/libraries/jquery-ui.less
@@ -1,3 +1,0 @@
-.ui-state-highlight {
-    height: 50px;
-}
