--- main/core/Resources/less/libraries/video-js.less
@@ -1,5 +1,0 @@
-// Avoid having the default player to appear and blink while video js is loading
-.vjs-waiting {
-    display: none;
-    background: transparent;
-}
