--- main/core/Resources/modules/scaffolding/ip/TODO
@@ -1,1 +1,0 @@
-move in IP data type module.
