+++ main/core/Resources/modules/validation/index.js
@@ -3,7 +3,6 @@ import set from 'lodash/set'
 import moment from 'moment'
 
 import {trans, tval} from '#/main/app/intl/translation'
-import {IPv4} from '#/main/core/scaffolding/ip'
 
 // TODO : break me
 
@@ -122,16 +121,6 @@ function lengthInRange(value, options) {
 function url(value) {
   if (match(value, {regex: /^((http[s]?|ftp):\/)?\/?([^:\/\s]+)((\/\w+)*\/)([\w\-\.]+[^#?\s]+)(.*)?(#[\w\-]+)?$/ig})) {
     return tval('This value should be a valid url.')
-  }
-}
-
-function ip(value) {
-  if (string(value)) {
-    return string(value)
-  }
-
-  if (!IPv4.isValid(value)) {
-    return tval('This value should be a valid IPv4.')
   }
 }
 
@@ -273,7 +262,6 @@ export {
   array,
   string,
   notBlank,
-  ip,
   match,
   number,
   inRange,
