+++ main/core/Resources/translations/platform.en.json
@@ -1582,7 +1582,9 @@
     "main_organization": "Main organization",
     "vat_number": "Vat number",
     "open_url": "Open url",
-    "resource_allowed_ip": "Allowed IPs",
+    "allowed_ips": "Allowed IPs",
+    "add_ip": "Add an IP address",
+    "no_allowed_ip": "No IP is allowed.",
     "summary": "Summary",
     "overview": "Overview",
     "poster": "Poster",
