+++ main/core/Resources/translations/platform.fr.json
@@ -1617,7 +1617,9 @@
     "main_organization": "Organisation principale",
     "vat_number": "Numéro de TVA",
     "open_url": "Url d'ouverture",
-    "resource_allowed_ip": "IPs authorisées",
+    "allowed_ips": "IPs authorisées",
+    "add_ip": "Ajouter une adresse IP",
+    "no_allowed_ip": "Aucune IP n'est authorisée.",
     "summary": "Sommaire",
     "overview": "Page d'accueil",
     "poster": "Image de couverture",
