+++ main/core/Resources/translations/resource.fr.json
@@ -21,7 +21,6 @@
     "resource_explorer": "Explorateur de ressources",
     "resource_fullscreen": "Afficher en plein écran à l'ouverture",
     "resource_likes": "{0} Cette ressource n'est aimée par aucun utilisateur. | {1} Cette ressource est aimée par un utilisateur. | ]1,Inf[ Cette ressource est aimée par %count% utilisateurs.",
-    "resource_no_allowed_ip": "Aucune IP n'est authorisée.",
     "resource_not_published": "La ressource n'est pas publiée",
     "resource_overview": "Vue d'ensemble de la ressource",
     "resource_overview_actions": "Actions disponibles",
